Introduces the stack create command, moves install command to be an alias, updates README and bumps version
