+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tembo"
-version = "0.3.0"
+version = "0.4.0"
 edition = "2021"
 authors = ["Tembo.io"]
 description = "The CLI for Tembo"
