+++ src/cli/mod.rs
@@ -1,3 +1,3 @@
 pub mod config;
 pub mod docker;
-pub mod stack;
+pub mod stacks;
