+++ src/cmd/mod.rs
@@ -1,2 +1,3 @@
 pub mod init;
 pub mod install;
+pub mod stack;
